Add Zustand caching to admin modules
