+++ app/store/use-api-cache.js
@@ -1,3 +1,4 @@
+"use client";
 import { create } from 'zustand'
 
 export const useApiCache = create((set, get) => ({
