+++ components/categoryTable.jsx
@@ -38,6 +38,7 @@ import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { toast } from "sonner";
 import apiFetch from "@/helpers/apiFetch";
+import { useApiCache } from "@/app/store/use-api-cache";
 
 function CategoryActions({ category, canEdit }) {
   const router = useRouter();
@@ -53,6 +54,8 @@ function CategoryActions({ category, canEdit }) {
       const data = await res.json();
       if (data.success) {
         toast.success('Status updated');
+        const clearData = useApiCache.getState().clearData;
+        clearData('categories');
         router.refresh();
       } else {
         toast.error(data.error || 'Failed to update status');
